Moved timeout to FunctionDispatcher and updated tests
